fixing test to poll_instance_tag_value_spec :tada:
