+++ aws_helpers.gemspec
@@ -37,4 +37,5 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency 'coveralls'
   spec.add_development_dependency 'rdiscount'
   spec.add_development_dependency 'codeclimate-test-reporter'
+  spec.add_development_dependency "recursive-open-struct",  "= 1.0.1"
 end
